requirements.txt
amqp==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.1
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crypto==1.4.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dependency-injector==4.38.0
dill==0.3.7
distlib==0.3.7
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.12.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
Naked==0.1.32
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pipenv==2023.11.15
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
starlette==0.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
